NIFI-10251 addressing additional review comments, cleanup: - use get instead of whenComplete because whenComplete is not blocking - prevent stucked processor when subscribe fails, yield on any exception - merge client and connection properties - create actual client in adapter, this way clientProperties will be used only in the adapter - remove timeout from disconnect signature because it is not applicable to all clients - add javadoc
