+++ nifi-nar-bundles/nifi-mqtt-bundle/nifi-mqtt-processors/src/main/java/org/apache/nifi/processors/mqtt/common/ReceivedMqttMessage.java
@@ -16,6 +16,9 @@
  */
 package org.apache.nifi.processors.mqtt.common;
 
+/**
+ * Represents a received MQTT message
+ */
 public class ReceivedMqttMessage extends StandardMqttMessage {
 
     private String topic;
