+++ nifi-nar-bundles/nifi-mqtt-bundle/nifi-mqtt-processors/src/main/java/org/apache/nifi/processors/mqtt/common/StandardMqttMessage.java
@@ -16,6 +16,9 @@
  */
 package org.apache.nifi.processors.mqtt.common;
 
+/**
+ * Represents a MQTT message.
+ */
 public class StandardMqttMessage {
     private byte[] payload;
     private int qos;
